AI System is functioning, test is in main.cpp
Several classes are now broken out into .h/.cpp file pairs. For some
reason, the main AISystem class didn't work this way so i left it as one
header file. The AISystem header now just includes the generic enemy
behavior class, and in main, all that is required is to include the
AISystem.h file and you're good to go. there were several changes I had
to make to get everything to work correctly, but for the most part it
all looks to be in working order.